src/third_party/blink/web_tests/display-lock/lock-before-append/locked-shadow-descendant.html

<!doctype HTML>

<div id="host">
  <input id="slotted" type="text"/>
</div>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
async_test((t) => {
  let container = document.createElement("div");
  container.id = "container";
  container.style = "contain: style layout;";
  container.innerHTML = "<slot></slot>";

  let shadowRoot = host.attachShadow({ mode: "open" });

  async function focusTest() {
    t.step(() => assert_not_equals(document.activeElement, slotted));
    t.step(() => assert_not_equals(shadowRoot.activeElement, slotted));
    t.step(() => assert_not_equals(container.displayLock, null));

    let acquire_promise = container.displayLock.acquire({ timeout: Infinity });

    await acquire_promise;
    shadowRoot.appendChild(container);

    t.step(() => assert_not_equals(document.activeElement, slotted));
    t.step(() => assert_not_equals(shadowRoot.activeElement, slotted));

    slotted.focus();
    t.step(() => assert_not_equals(document.activeElement, slotted));
    t.step(() => assert_not_equals(shadowRoot.activeElement, slotted));

    forceLayoutTest();
  }

  async function forceLayoutTest() {
    t.step(() => assert_equals(slotted.offsetTop, 0));
    // Add a 20px div above the slotted div.
    container.innerHTML = "<div style='height: 20px;'></div><slot></slot>";
    t.step(() => assert_equals(slotted.offsetTop, 20));
    t.done();
  }

  window.onload = function() {
    requestAnimationFrame(() => requestAnimationFrame(focusTest));
  };
}, "Testing focus and force layout on element with locked flat-tree ancestor");

</script>
